requirements.txt
pyaml
docopt==0.6.2
gitpython
